backend/pyproject.toml
[project]
name = "bilihelper-backend"
version = "0.1.0"
description = "BiliHelper backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "passlib[argon2]>=1.7",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "cryptography>=43",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
